# salo-stac-api/openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  contact:
    name: STAC Specification
    url: http://stacspec.org
  description: This is an OpenAPI definition of the core SpatioTemporal Asset Catalog
    API specification. Any service that implements this endpoint to allow search of
    spatiotemporal assets can be considered a STAC API. The endpoint is also available
    as an OpenAPI fragment that can be integrated with other OpenAPI definitions,
    and is designed to slot seamlessly into a OGC API - Features definition.
  license:
    name: Apache License 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  title: The SpatioTemporal Asset Catalog API + Extensions
  version: 0.9.0
servers:
- description: Development server
  url: http://dev.cool-sat.com
- description: Production server
  url: http://www.cool-sat.com
...
paths:
...
  /collections/{collectionId}/items:
    get:
      description: |-
        Fetch features of the feature collection with id `collectionId`.

        Every feature in a dataset belongs to a collection. A dataset may
        consist of multiple feature collections. A feature collection is often a
        collection of features of a similar type, based on a common schema.

        Use content negotiation to request HTML or GeoJSON.
      operationId: get_features
      parameters:
      - description: local identifier of a collection
        explode: false
        in: path
        name: collectionId
        required: true
        schema:
          type: string
        style: simple
      - description: |-
          The optional limit parameter limits the number of items that are presented in the response document.

          Only items are counted that are on the first level of the collection in the response document.
          Nested objects contained within the explicitly requested items shall not be counted.

          Minimum = 1. Maximum = 10000. Default = 10.
        explode: false
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 10000
          minimum: 1
          type: integer
        style: form
      - description: |-
          Only features that have a geometry that intersects the bounding box are selected.
          The bounding box is provided as four or six numbers, depending on
          whether the coordinate reference system includes a vertical axis (height
          or depth):

          * Lower left corner, coordinate axis 1
          * Lower left corner, coordinate axis 2
          * Minimum value, coordinate axis 3 (optional)
          * Upper right corner, coordinate axis 1
          * Upper right corner, coordinate axis 2
          * Maximum value, coordinate axis 3 (optional)

          The coordinate reference system of the values is WGS 84
          longitude/latitude (http://www.opengis.net/def/crs/OGC/1.3/CRS84) unless
          a different coordinate reference system is specified in the parameter
          `bbox-crs`.

          For WGS 84 longitude/latitude the values are in most cases the sequence
          of minimum longitude, minimum latitude, maximum longitude and maximum
          latitude. However, in cases where the box spans the antimeridian the
          first value (west-most box edge) is larger than the third value
          (east-most box edge).

          If the vertical axis is included, the third and the sixth number are
          the bottom and the top of the 3-dimensional bounding box.

          If a feature has multiple spatial geometry properties, it is the
          decision of the server whether only a single spatial geometry property
          is used to determine the extent or all relevant geometries.

          Example: The bounding box of the New Zealand Exclusive Economic Zone in
          WGS 84 (from 160.6°E to 170°W and from 55.95°S to 25.89°S) would be
          represented in JSON as `[160.6, -55.95, -170, -25.89]` and in a query as
          `bbox=160.6,-55.95,-170,-25.89`.
        explode: false
        in: query
        name: bbox
        required: false
        schema:
          items:
            type: number
          maxItems: 6
          minItems: 4
          type: array
        style: form
      - description: |-
          Either a date-time or an interval, open or closed. Date and time expressions
          adhere to RFC 3339. Open intervals are expressed using double-dots.

          Examples:

          * A date-time: "2018-02-12T23:20:50Z"
          * A closed interval: "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"
          * Open intervals: "2018-02-12T00:00:00Z/.." or "../2018-03-18T12:31:12Z"

          Only features that have a temporal property that intersects the value of
          `datetime` are selected.

          If a feature has multiple temporal properties, it is the decision of the
          server whether only a single temporal property is used to determine
          the extent or all relevant temporal properties.
        explode: false
        in: query
        name: datetime
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/geo+json:
              schema:
                $ref: '#/components/schemas/featureCollectionGeoJSON'
            text/html:
              schema:
                type: string
          description: |-
            The response is a document consisting of features in the collection.
            The features included in the response are determined by the server
            based on the query parameters of the request. To support access to
            larger collections without overloading the client, the API supports
            paged access with links to the next page, if more features are selected
            that the page size.

            The `bbox` and `datetime` parameter can be used to select only a
            subset of the features in the collection (the features that are in the
            bounding box or time interval). The `bbox` parameter matches all features
            in the collection that are not associated with a location, too. The
            `datetime` parameter matches all features in the collection that are
            not associated with a time stamp or interval, too.

            The `limit` parameter may be used to control the subset of the
            selected features that should be returned in the response, the page size.
            Each page may include information about the number of selected and
            returned features (`numberMatched` and `numberReturned`) as well as
            links to support paging (link relation `next`).
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A query parameter has an invalid value.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: fetch features
      tags:
      - Data
      x-openapi-router-controller: openapi_server.controllers.data_controller
    post:
      description: create a new feature in a specific collection
      operationId: post_feature
      parameters:
      - description: local identifier of a collection
        explode: false
        in: path
        name: collectionId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - $ref: '#/components/schemas/item'
              - $ref: '#/components/schemas/itemCollection'
      responses:
        "201":
          content:
            application/geo+json:
              schema:
                type: string
            text/html:
              schema:
                type: string
          description: Status of the create request.
          headers:
            Location:
              description: A link to the item
              explode: false
              schema:
                format: url
                type: string
              style: simple
            ETag:
              description: A string to ensure the item has not been modified
              explode: false
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was malformed or semantically invalid
        "5XX":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was syntactically and semantically valid, but an
            error occurred while trying to act upon it
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: An error occurred.
      summary: add a new feature to a collection
      tags:
      - Transaction Extension
      x-openapi-router-controller: openapi_server.controllers.transaction_extension_controller
  /collections/{collectionId}/items/{featureId}:
    delete:
      description: Use this method to delete an existing feature.
      operationId: delete_feature
      parameters:
      - description: local identifier of a collection
        explode: false
        in: path
        name: collectionId
        required: true
        schema:
          type: string
        style: simple
      - description: local identifier of a feature
        explode: false
        in: path
        name: featureId
        required: true
        schema:
          type: string
        style: simple
      - description: Only take the action if the ETag of the item still matches
        explode: false
        in: header
        name: If-Match
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Status of the delete request.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was malformed or semantically invalid
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The specified resource was not found
        "5XX":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was syntactically and semantically valid, but an
            error occurred while trying to act upon it
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: An error occurred.
      summary: delete an existing feature by Id
      tags:
      - Transaction Extension
      x-openapi-router-controller: openapi_server.controllers.transaction_extension_controller
    get:
      description: |-
        Fetch the feature with id `featureId` in the feature collection
        with id `collectionId`.

        Use content negotiation to request HTML or GeoJSON.
      operationId: get_feature
      parameters:
      - description: local identifier of a collection
        explode: false
        in: path
        name: collectionId
        required: true
        schema:
          type: string
        style: simple
      - description: local identifier of a feature
        explode: false
        in: path
        name: featureId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/geo+json:
              schema:
                $ref: '#/components/schemas/item'
            text/html:
              schema:
                type: string
          description: |-
            fetch the feature with id `featureId` in the feature collection
            with id `collectionId`
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: fetch a single feature
      tags:
      - Data
      x-openapi-router-controller: openapi_server.controllers.data_controller
    patch:
      description: Use this method to update an existing feature. Requires a GeoJSON
        fragement (containing the fields to be updated) be submitted.
      operationId: patch_feature
      parameters:
      - description: local identifier of a collection
        explode: false
        in: path
        name: collectionId
        required: true
        schema:
          type: string
        style: simple
      - description: local identifier of a feature
        explode: false
        in: path
        name: featureId
        required: true
        schema:
          type: string
        style: simple
      - description: Only take the action if the ETag of the item still matches
        explode: false
        in: header
        name: If-Match
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/partialItem'
      responses:
        "200":
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: string
          description: Status of the update request.
          headers:
            ETag:
              description: A string to ensure the item has not been modified
              explode: false
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was malformed or semantically invalid
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The specified resource was not found
        "5XX":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was syntactically and semantically valid, but an
            error occurred while trying to act upon it
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: An error occurred.
      summary: update an existing feature by Id with a partial item definition
      tags:
      - Transaction Extension
      x-openapi-router-controller: openapi_server.controllers.transaction_extension_controller
    put:
      description: Use this method to update an existing feature. Requires the entire
        GeoJSON  description be submitted.
      operationId: put_feature
      parameters:
      - description: local identifier of a collection
        explode: false
        in: path
        name: collectionId
        required: true
        schema:
          type: string
        style: simple
      - description: local identifier of a feature
        explode: false
        in: path
        name: featureId
        required: true
        schema:
          type: string
        style: simple
      - description: Only take the action if the ETag of the item still matches
        explode: false
        in: header
        name: If-Match
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/item'
      responses:
        "200":
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: string
          description: Status of the update request.
          headers:
            ETag:
              description: A string to ensure the item has not been modified
              explode: false
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was malformed or semantically invalid
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The specified resource was not found
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: Some condition specified by the request could not be met in
            the server
        "5XX":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
          description: The request was syntactically and semantically valid, but an
            error occurred while trying to act upon it
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: An error occurred.
      summary: update an existing feature by Id with a complete item definition
      tags:
      - Transaction Extension
      x-openapi-router-controller: openapi_server.controllers.transaction_extension_controller
...
components:
...
  schemas:
...
    exception:
      description: 'Information about the exception: an error code plus an optional
        description.'
      properties:
        code:
          type: string
        description:
          type: string
      required:
      - code
      type: object
...
    featureCollectionGeoJSON:
      example:
        timeStamp: 2017-08-17T08:05:32Z
        features:
        - stac_version: 0.9.0
          stac_extensions:
          - eo
          - view
          - https://example.com/cs-extension/1.0/schema.json
          type: Feature
          id: CS3-20160503_132131_05
          bbox:
          - -122.59750209
          - 37.48803556
          - -122.2880486
          - 37.613537207
          geometry:
            type: Polygon
            coordinates:
            - - - -122.308150179
                - 37.488035566
              - - -122.597502109
                - 37.538869539
              - - -122.576687533
                - 37.613537207
              - - -122.2880486
                - 37.562818007
              - - -122.308150179
                - 37.488035566
          properties:
            datetime: 2016-05-03T13:22:30.040Z
            title: A CS3 item
            license: PDDL-1.0
            providers:
            - name: CoolSat
              roles:
              - producer
              - licensor
              url: https://cool-sat.com/
            view:sun_azimuth: 168.7
            eo:cloud_cover: 0.12
            view:off_nadir: 1.4
            platform: coolsat2
            instruments:
            - cool_sensor_v1
            eo:bands: []
            view:sun_elevation: 33.4
            eo:gsd: 0.512
          collection: CS3
          links:
          - rel: self
            href: http://cool-sat.com/collections/CS3/items/20160503_132130_04
          - rel: root
            href: http://cool-sat.com/collections
          - rel: parent
            href: http://cool-sat.com/collections/CS3
          - rel: collection
            href: http://cool-sat.com/collections/CS3
          assets:
            analytic:
              href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/analytic.tif
              title: 4-Band Analytic
            thumbnail:
              href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/thumbnail.png
              title: Thumbnail
        - stac_version: 0.9.0
          stac_extensions:
          - eo
          - view
          - https://example.com/cs-extension/1.0/schema.json
          type: Feature
          id: CS3-20160503_132131_05
          bbox:
          - -122.59750209
          - 37.48803556
          - -122.2880486
          - 37.613537207
          geometry:
            type: Polygon
            coordinates:
            - - - -122.308150179
                - 37.488035566
              - - -122.597502109
                - 37.538869539
              - - -122.576687533
                - 37.613537207
              - - -122.2880486
                - 37.562818007
              - - -122.308150179
                - 37.488035566
          properties:
            datetime: 2016-05-03T13:22:30.040Z
            title: A CS3 item
            license: PDDL-1.0
            providers:
            - name: CoolSat
              roles:
              - producer
              - licensor
              url: https://cool-sat.com/
            view:sun_azimuth: 168.7
            eo:cloud_cover: 0.12
            view:off_nadir: 1.4
            platform: coolsat2
            instruments:
            - cool_sensor_v1
            eo:bands: []
            view:sun_elevation: 33.4
            eo:gsd: 0.512
          collection: CS3
          links:
          - rel: self
            href: http://cool-sat.com/collections/CS3/items/20160503_132130_04
          - rel: root
            href: http://cool-sat.com/collections
          - rel: parent
            href: http://cool-sat.com/collections/CS3
          - rel: collection
            href: http://cool-sat.com/collections/CS3
          assets:
            analytic:
              href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/analytic.tif
              title: 4-Band Analytic
            thumbnail:
              href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/thumbnail.png
              title: Thumbnail
        numberReturned: 10
        links:
        - headers:
            Accept: application/json
          hreflang: en
          method: GET
          merge: false
          rel: child
          length: 0
          href: http://www.geoserver.example/stac/naip/child/catalog.json
          type: application/geo+json
          title: NAIP Child Catalog
          body: '{}'
        - headers:
            Accept: application/json
          hreflang: en
          method: GET
          merge: false
          rel: child
          length: 0
          href: http://www.geoserver.example/stac/naip/child/catalog.json
          type: application/geo+json
          title: NAIP Child Catalog
          body: '{}'
        type: FeatureCollection
        numberMatched: 127
      properties:
        type:
          enum:
          - FeatureCollection
          type: string
        features:
          items:
            $ref: '#/components/schemas/item'
          type: array
        links:
          items:
            $ref: '#/components/schemas/link'
          type: array
        timeStamp:
          description: This property indicates the time and date when the response
            was generated.
          example: 2017-08-17T08:05:32Z
          format: date-time
          type: string
        numberMatched:
          description: |-
            The number of features of the feature type that match the selection
            parameters like `bbox`.
          example: 127
          minimum: 0
          type: integer
        numberReturned:
          description: |-
            The number of features in the feature collection.

            A server may omit this information in a response, if the information
            about the number of features is not known or difficult to compute.

            If the value is provided, the value shall be identical to the number
            of items in the "features" array.
          example: 10
          minimum: 0
          type: integer
      required:
      - features
      - type
      type: object
...
    item:
      description: A GeoJSON Feature augmented with foreign members that contain values
        relevant to a STAC entity
      example:
        stac_version: 0.9.0
        stac_extensions:
        - eo
        - view
        - https://example.com/cs-extension/1.0/schema.json
        type: Feature
        id: CS3-20160503_132131_05
        bbox:
        - -122.59750209
        - 37.48803556
        - -122.2880486
        - 37.613537207
        geometry:
          type: Polygon
          coordinates:
          - - - -122.308150179
              - 37.488035566
            - - -122.597502109
              - 37.538869539
            - - -122.576687533
              - 37.613537207
            - - -122.2880486
              - 37.562818007
            - - -122.308150179
              - 37.488035566
        properties:
          datetime: 2016-05-03T13:22:30.040Z
          title: A CS3 item
          license: PDDL-1.0
          providers:
          - name: CoolSat
            roles:
            - producer
            - licensor
            url: https://cool-sat.com/
          view:sun_azimuth: 168.7
          eo:cloud_cover: 0.12
          view:off_nadir: 1.4
          platform: coolsat2
          instruments:
          - cool_sensor_v1
          eo:bands: []
          view:sun_elevation: 33.4
          eo:gsd: 0.512
        collection: CS3
        links:
        - rel: self
          href: http://cool-sat.com/collections/CS3/items/20160503_132130_04
        - rel: root
          href: http://cool-sat.com/collections
        - rel: parent
          href: http://cool-sat.com/collections/CS3
        - rel: collection
          href: http://cool-sat.com/collections/CS3
        assets:
          analytic:
            href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/analytic.tif
            title: 4-Band Analytic
          thumbnail:
            href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/thumbnail.png
            title: Thumbnail
      properties:
        stac_version:
          example: 0.9.0
          title: STAC version
          type: string
        stac_extensions:
          items:
            anyOf:
            - format: uri
              title: Reference to a JSON Schema
              type: string
            - title: Reference to a core extension
              type: string
          title: STAC extensions
          type: array
          uniqueItems: true
        id:
          description: Provider identifier, a unique ID, potentially a link to a file.
          example: path/to/example.tif
          type: string
        bbox:
          description: "Only features that have a geometry that intersects the bounding\
            \ box are\nselected. The bounding box is provided as four or six numbers,\n\
            depending on whether the coordinate reference system includes a\nvertical\
            \ axis (elevation or depth):\n\n* Lower left corner, coordinate axis 1\n\
            * Lower left corner, coordinate axis 2  \n* Lower left corner, coordinate\
            \ axis 3 (optional) \n* Upper right corner, coordinate axis 1 \n* Upper\
            \ right corner, coordinate axis 2 \n* Upper right corner, coordinate axis\
            \ 3 (optional)\n\nThe coordinate reference system of the values is WGS84\n\
            longitude/latitude (http://www.opengis.net/def/crs/OGC/1.3/CRS84) unless\n\
            a different coordinate reference system is specified in the parameter\n\
            `bbox-crs`.\n\nFor WGS84 longitude/latitude the values are in most cases\
            \ the sequence\nof minimum longitude, minimum latitude, maximum longitude\
            \ and maximum\nlatitude. However, in cases where the box spans the antimeridian\
            \ the\nfirst value (west-most box edge) is larger than the third value\n\
            (east-most box edge).\n\nIf a feature has multiple spatial geometry properties,\
            \ it is the\ndecision of the server whether only a single spatial geometry\
            \ property\nis used to determine the extent or all relevant geometries.\n\
            \nExample: The bounding box of the New Zealand Exclusive Economic Zone\
            \ in\nWGS 84 (from 160.6°E to 170°W and from 55.95°S to 25.89°S) would\
            \ be\nrepresented in JSON as `[160.6, -55.95, -170, -25.89]` and in a\
            \ query as\n`bbox=160.6,-55.95,-170,-25.89`."
          example:
          - -110
          - 39.5
          - -105
          - 40.5
          items:
            type: number
          maxItems: 6
          minItems: 4
          type: array
        geometry:
          $ref: '#/components/schemas/Geometry'
        type:
          $ref: '#/components/schemas/itemType'
        properties:
          $ref: '#/components/schemas/itemProperties'
        links:
          items:
            $ref: '#/components/schemas/link'
          type: array
        assets:
          additionalProperties:
            properties:
              href:
                description: Link to the asset object
                example: http://cool-sat.com/catalog/collections/cs/items/CS3-20160503_132130_04/thumb.png
                format: url
                type: string
              title:
                description: Displayed title
                example: Thumbnail
                type: string
              description:
                description: |-
                  Multi-line description to explain the asset.

                  [CommonMark 0.29](http://commonmark.org/) syntax MAY be used for
                  rich text representation.
                example: Small 256x256px PNG thumbnail for a preview.
                type: string
              type:
                description: Media type of the asset
                example: image/png
                type: string
              roles:
                description: Purposes of the asset
                example:
                - thumbnail
                items:
                  type: string
                type: array
            required:
            - href
            type: object
          type: object
      required:
      - assets
      - bbox
      - geometry
      - id
      - links
      - properties
      - stac_version
      - type
      type: object
...
    itemType:
      description: The GeoJSON type
      enum:
      - Feature
      type: string
...
    Geometry:
      oneOf:
      - $ref: '#/components/schemas/GeoJSON_Point'
      - $ref: '#/components/schemas/GeoJSON_LineString'
      - $ref: '#/components/schemas/GeoJSON_Polygon'
      - $ref: '#/components/schemas/GeoJSON_MultiPoint'
      - $ref: '#/components/schemas/GeoJSON_MultiLineString'
      - $ref: '#/components/schemas/GeoJSON_MultiPolygon'
      title: GeoJSON Geometry
...
    GeoJSON_Point:
      properties:
        type:
          enum:
          - Point
          type: string
        coordinates:
          items:
            type: number
          minItems: 2
          type: array
        bbox:
          items:
            type: number
          minItems: 4
          type: array
      required:
      - coordinates
      - type
    GeoJSON_LineString:
      properties:
        type:
          enum:
          - LineString
          type: string
        coordinates:
          items:
            items:
              type: number
            minItems: 2
            type: array
          minItems: 2
          type: array
        bbox:
          items:
            type: number
          minItems: 4
          type: array
      required:
      - coordinates
      - type
...